done + css

+++ Django_Guestbook/static/js/guestbook/widget/GreetingWidget.js
@@ -18,10 +18,11 @@ define([
     "dijit/form/ValidationTextBox",
     "dijit/InlineEditBox",
     "dijit/form/Textarea",
+    "dojo/fx/Toggler",
     "dojo/text!./templates/GreetingWidget.html"
 ], function(request, parser, ready, cookie, declare, baseFx, lang, domStyle, mouse, on,
             GuestbookWidget, _WidgetBase, _OnDijitClickMixin, _TemplatedMixin,_WidgetsInTemplateMixin,
-            Button, ValidationTextBox, InlineEditBox, Textarea, template){
+            Button, ValidationTextBox, InlineEditBox, Textarea, Toggler, template){
     return declare([_WidgetBase,_OnDijitClickMixin, _TemplatedMixin, _WidgetsInTemplateMixin], {
         author: "An anonymous",
         content: "",
@@ -37,15 +38,9 @@ define([
         guestbook_name : "",
         is_author : false,
         id_greeting : "",
-
+        disabled: "",
+        hidden : "none",
         postCreate: function(){
-            this.deleteButtonNode.disabled = true;
-            this.contentNode.disabled = true;
-            if (this.is_admin)
-                this.deleteButtonNode.disabled = false;
-                this.contentNode.disabled = false;
-            if (this.is_author)
-                this.contentNode.disabled = false;
             var domNode = this.domNode;
             this.inherited(arguments);
             domStyle.set(domNode, "backgroundColor", this.baseBackgroundColor);
@@ -56,6 +51,21 @@ define([
                 on(domNode, mouse.leave, lang.hitch(this, "_changeBackground", this.baseBackgroundColor))
             );
         },
+
+        buildRendering: function(){
+            if(this.is_admin){
+                this.hidden = "display: true";
+            }
+            else
+                this.hidden = "display: none";
+            if(this.is_admin || this.is_author){
+                this.disabled = "disabled: false,"
+            }
+            else
+                this.disabled = "disabled: true,";
+            this.inherited(arguments);
+        },
+
         startup: function(){
             this.inherited(arguments);
         },
@@ -75,29 +85,41 @@ define([
         },
 
         _delete: function(guestbook, id){
+            if(!this.is_admin){
+                alert("You are not administrator!!!");
+            }
+            else{
+                request.del("/api/guestbook/"+guestbook+"/greeting/"+id+"/", {
+                    headers:{
+                        "X-CSRFToken": cookie('csrftoken')
+                    },
+                    timeout : 1000
+                });
+                dijit.byId("guestbook")._loadgreeting(guestbook, 500);
+            }
 
-            request.del("/api/guestbook/"+guestbook+"/greeting/"+id+"/", {
-                headers:{
-                    "X-CSRFToken": cookie('csrftoken')
-                },
-                timeout : 1000
-            });
-            dijit.byId("guestbook")._loadgreeting(guestbook, 500);
 
         },
 
         _put: function(guestbook, id){
             //this.contentNode.value
-            request.put("/api/guestbook/"+guestbook+"/greeting/"+id+"/", {
-                data:{
-                    content: this.contentNode.value
-                },
-                headers:{
-                    "X-CSRFToken": cookie('csrftoken')
-                },
-                timeout : 1000
-            });
-            dijit.byId("guestbook")._loadgreeting(guestbook, 500);
+            if(this.is_admin || this.is_author){
+                request.put("/api/guestbook/"+guestbook+"/greeting/"+id+"/", {
+                    data:{
+                        content: this.contentNode.value
+                    },
+                    headers:{
+                        "X-CSRFToken": cookie('csrftoken')
+                    },
+                    timeout : 1000
+                });
+                dijit.byId("guestbook")._loadgreeting(guestbook, 500);
+
+            }
+            else{
+                alert("You don't have permisson to update this greeting!!")
+            }
+
         }
 
     });
